requirements.txt
upstox-python-sdk
websocket-client
pandas
numpy
requests
